% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_plot.R
\name{empty_plot}
\alias{empty_plot}
\title{empty_pot}
\usage{
empty_plot(label = "No valid data to generate a plot")
}
\arguments{
\item{label}{Text that should be returned in the center of the plot}
}
\value{
ggplot2 plot with labeled text
}
\description{
empty_pot
}
