% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTplotapi.R
\name{plot_interactive}
\alias{plot_interactive}
\title{In Vivo Toxicity Scatter Plot by exposure route}
\arguments{
\item{exposure_route}{"oral" or "inhalation"}

\item{predicted_exposure}{predicted BER exposure}

\item{title}{plot title}
}
\description{
In Vivo Toxicity Scatter Plot by exposure route
}
