% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTplot.R
\name{RTplot}
\alias{RTplot}
\title{RT plot}
\usage{
RTplot(
  dat,
  exposure_route,
  predicted_exposure,
  title,
  xaxis_label = "Exposure Duration Class"
)
}
\arguments{
\item{dat}{dataframe with plotting parameters}

\item{exposure_route}{"oral" or "inhalation"}

\item{predicted_exposure}{predicted BER exposure}

\item{title}{plot title}

\item{xaxis_label}{String with default "Exposure Duration Class" can be changed to change the xaxis}
}
\value{
RT plot
}
\description{
RT plot
}
